Fix system.options.boot overwrite
The function get-current-exec is supposed to out-prioritize
the argv[0] contents, which may not be correct.  (GitHub
Actions seems to have one of those cases where it's not
correct, e.g. on Ubuntu)

However it was unconditionally overwriting what the
get-current-exec gave with argv[0], due to the absence of a
use of DEFAULT.  This fixes that.