Void handling in ATTEMPT
The ATTEMPT function is built on top of REDUCE-EACH were
every step is evaluated ^META, so that definitional errors
can be reacted to.

However, it was returning the UNMETA of whatever the last
reduce step produced.  That was inside a branch, so a void
would be transformed to a none...yet overwriting any result
prior to the void.

This rethinks it so that an accumulator for the last value
is kept, and the accumulator is not overwritten on void
steps.  The accumulator is still initialized to none, so
that void is not returned by ATTEMPT (which would be conflated
by ELSE as something to react to, when the expectation is
that ELSE would be something that ran on failure).