REDUCE decisions/fixes
This commits to the idea that REDUCE does not offer voids
to predicates, and that void filtering is part of the reusable
work that REDUCE is supposed to do that makes a predicate
worthwhile.

Those who want more control can use REDUCE-EACH with a ^META
variable.

For predicates that take voidable ^META parameters, this *may*
be revisited in the future...but it's a somewhat fringe use.