Partial resurrection of MACRO for stackless
MACRO and INLINE previously used the implicit retriggering
of the evaluator on pure invisibility to accomplish their
effects.  This retriggering no longer exists, so if the
macro splices more code into the feed stream and then
returns a void, the overall macro evaluates to a void that
things taking arguments would see.

This moves the retriggering into the INLINE function and
macro dispatcher, by pushing an evaluator frame.  This does
not work for enfix...so that will have to be rethought if
it is to be supported.