Update bad decompression logic in LOAD
R3-Alpha compressed scripts either with base64-encoding (as
a base64 BINARY! literal) or with raw compression bytes.  It
did not indicate which was used in the header, so it used
ATTEMPT to do a raw decompression of what would be garbage...
and checked if that succeeded or not.

Trying to decompress garbage typically leads to a first step
of allocating a buffer sized based on some bytes in the garbage
which can be arbitrary, hence it could fail on an actual
attempt at a nonsensical memory allocation.

(The new semantics of ATTEMPT caught this memory allocation
failure, by not glossing over it.)

This wasn't tended to immediately to figure out what was wrong
because this script compression feature is not in use.  But
looking into it, clearly the header should be written in SAVE
to say which kind of decompression it is, vs. resorting to
such a method.

This is being tended to as part of an "unimportant broken
tests triage", so rather than implement a better version it
just documents it and changes it to TRAP.  But a rewrite
should change the method, as well as review other issues
related to script compression (including "is this necessary
to be doing at all").