Simplify const_if_c use in DEBUG_UTF8_EVERYWHERE
A trick using a macro called `const_if_c` was repeated in
the DEBUG_UTF8_EVERYWHERE code, but was using a different
name due to some complexities of the Utf8(*) definition
across the debug modes.

That code had been simplified, to where the DEBUG_UTF8 switch
was really only controlling the use of a class that stopped
direct incrementation of the byte pointers or dereferencing.

With that simplification, the const_if_c could be used in
the normal way...removing the need for the obfuscated name.