Restore PROTECT/HIDE feature
The PROTECT command was restyled in terms of the fledgling
new behavior behind "pathing" (e.g. "tupling"), to navigate
down a TUPLE! and find the fields it corresponds to.

It used the same method as SET to find the field to update,
but at the last step would call PROTECT* instead of POKE* to
do the cell updating.

However, it was only equipped with a LOGIC! to do the update
of the bits.  This changes the low-level protect to take
one of 3 instructions: PROTECT, UNPROTECT, or HIDE.

This restores the historical ability to hide object fields
(which is a questionable feature, but kind of interesting.)