New INTEGER! parse semantics, SKIP combinator
The historical idea in Redbol parse was that INTEGER!s would
act as a prefix operator dictating a repeat count.  But that
led to some contortions, like SKIP being an arity-0 operator
where if you wanted to skip 3 things you would say [3 skip]

It seems like a better idea to make the integer just evaluate
to itself as a combinator.  Then if you want to repeat, you
can do so explicitly with REPEAT 3 RULE instead of 3 RULE.
And SKIP can take a parameter of how much to skip, as with
SKIP 3.

This makes that semantic change.  It affects a lot of tests,
but tests seem to be the place where the hardcoded number of
times to repeat a rule appears most often.  (It does not come
up that often in regular code.)