Remove support for FUNC without RETURN:
A very old behavior of definitional-return FUNCs was that
if you happened to call a refinement or parameter in the
spec of a function /RETURN, then it would let that preempt
the creation of a definitional return.

This no longer makes sense, because RETURN is the only way
to get a value back from something declaring itself a
function.

LAMBDA is now the approved way to create a function-like
thing which has no RETURN.