Raise errors in BLOCK! combinator
The ^META error machinery is still being designed.  And as
written it has a bit of an issue where if something takes
a ^META return value--gets an ERROR!--and then writes over
it, the error it represents will be lost.

For the moment this makes the BLOCK! combinator the catch
all to raise errors if not intercepted by any of the other
combinators.  Best practices are not yet understood for this
mechanism--as it is novel--so it will all have to be revisited.