Test IS_NATIVE flag on archetype
The ability to HIJACK actions and put a different archetype
in the identity (details) means that the flags may not match
on the archetype and identity.

However, when interpreting the details array as to whether
it is for a native or not, that has to be done from info
on the details array itself...which is the action identity.