Implement WANTED? by testing adjacent frame cell
The original idea behind knowing whether a multi-return was
"wanted" by the caller or not presumed a single cell in the
frame for the output, and that the initial state of that
cell would communicate whether the caller wanted the value
(by being ~wanted~ or a ~ isotope)

But the most obvious implementation of multi-returns turned
out to actually store the variable to write back in a hidden
local slot next to the named output.  This means the WANTED?
question can be answered by looking at the contents of that
adjacent cell.

Multi-returns are in flux as to whether they will be a "high
level" service understood by the basic frame mechanics, or
something that natives and FUNCs add on that isn't present
in the lower-level lambdas.  So this could create some problems
if WANTED? can't make assumptions about what kind of generator
made the multi-return... and necessitate "definitional WANTED?"
where the generator provides it.

In any case, ~wanted~ as a reified isotope was not a good
solution to the problem...so going with a hardcoded WANTED?
that can communicate with natives and FUNC based on the hidden
variable convention is an improvement for now.