Patch behavior for protect of WORD!
PICKing and POKEing were unified with TUPLE! dispatch, where
a tuple is picked in steps with a POKE run on the final step.
Other code that might want to fiddle with a value reached
through tuple steps reuses the machinery and just runs a
different last step.

PROTECT is an example of such code that runs an "updater"
in the final step.  However, the UPDATER was not being called
in the case of a WORD!...it was using a SET of the variable.
That works for POKE but not when the intent was to PROTECT.

This wedges in some behavior for that, which needs to be
revisited as this experimental methodology solidifies.