F.RETURN vs. UNWIND in parse STOP/BREAK
UNWIND is a generic and kind of dangerous operation, which
bypasses any type checking or cleanup that is done by a
combinator.  As a notable problem, this would bypass the
writeback of multi-return data...so even if a command like
stop tries to set the remainder to the input for a loop,
that would not be written back.

Since we're already making assumptions about the frame (like
that it has a .remainder to be set), we can go ahead and
call the F.RETURN in that frame, instead of resorting to
UNWIND (which should probably be called SYS.UTIL.UNWIND or
otherwise hidden).