Move multi-return proxying to dispatchers
Multi-returns are kind of on the fence now as partially a
core system behavior of functions, and potentially something
offered only by certain dispatchers.

Having the code to proxy multi-returns be something all
actions did was fundamentally breaking the ability to run a
copy of a frame in an ENCLOSE.  Fixing the immediate issue
by moving proxying into the relevant dispatchers introduces
another problem: that the enclose does not have the ability
to tweak the multi-return outputs, because the writeback to
the variables has already happened by the time the ENCLOSE
regains control after the DO.

This chooses to err on the side of not giving ENCLOSE access
to multi-returns (and having to cope with whatever the DO of
the FRAME! does with them).  This calls out the natives that
need proxying explicitly, and stops a current breakage in
Redbol, while causing a breakage in the trackparse demo.

More thought is needed--but changing in order to be able to
work on a Redbol issue.