Fix binding bug for all-byte tuples
The somewhat superfluous-seeming feature of packing numeric
tuples and paths as bytes was preserved.

Consequently a tuple like 0.0.0 does not have an underlying
array.  However one case in the binding code was attempting
to treat it like one.

This uses the correct tests with ANY_ARRAYLIKE/ANY_WORDLIKE
to ask about the underlying implementation in binding, vs the
surface type.