Bitset negation exception for EXCLUDE
R3-Alpha's BITSET! was based on BINARY! with an optional
"negated" flag.  This is an easy-enough way to to work with
bitsets if you never perform operations on them when they
are in the negated states--but if you try to do set operations
it becomes much more complicated.

There is a branch to deal with this that uses a C library
called "Roaring Bitsets".  But it's a fairly involved change
that hasn't been a priority to incorporate, relative to other
things that need to be thought about.

However, the %pdf-maker.r Rebol2 sample does a fairly simple
EXCLUDE operation which can be accommodated.