Turn LOOP back into an error
For a brief time, it was considered that WHILE would be arity-1
and LOOP would take it's place as the arity-2 "loop while true"
This was in order to get consistency with the parse construct
of arity-1 WHILE.

This was unsatisfying, and other tradeoffs were made in PARSE
to solve mitigate the elimination of ANY and WHILE as arity-1
looping operators.

Hence LOOP is put back in its role as the word reserved for
a full-featured "loop dialect", if that could be designed.