Change `<blank>` tag to `<try>`, bootstrap part
This makes substantial changes to the bootstrap code in
order to accommodate NULL as the no-op for APPEND/etc.
instead of BLANK!.

Making this somewhat tricky is that the bootstrap build will
not fetch NULL variables without a GET-WORD! or GET.  So
having refinements be null when unused means you can no longer
write `if ref` in the bootstrap code.  It would have to be
`if :ref` or `if get 'ref`.

Instead of using `if :ref`, this uses `if try ref` in order
to be more suggestive of having something to do with NULL
(as opposed to defusing functions).  Then TRY is retrofitted
in the bootstrap shim to be a quoting variadic that will
defuse the error on null.

This means there are TRY instances that are superfluous in
the modern build.  But it is working.