Changes for elimination of /ONLY, add SPREAD
The /ONLY refinement had been deprecated for some time, with
the idea that for operations like APPEND that it would be
the value getting appended that would carry the signal of
whether it should be spliced or not.

This idea had it that BLOCK! at quote level 0 would splice,
but at quote level > 0 would drop a quote level and not splice.
Then many types (such as WORD!) were disallowed from being
used at quote level 0, requiring to be quoted for safety.

Over time it became evident that a better idea would be to
let quote level 0 be "as is" and allow BLOCK! to have an
isotopic "quote level -1" as the splicing signal.  This
requires APPEND & friends to take the value as a ^META
argument, but the interpretation of the signals was already
basically having to do a "meta convention" behind the scenes.
This makes it official.

Massive changes are required to support this.  While it would
be more judicious to inventory and add them one by one, the
choice here is to go ahead with the pain of pushing it through
to see the impacts...