Rethink JOIN to use SPREAD
This simplifies JOIN some, by making it follow more directly
with the idea of acting like APPEND to a COPY of the input.

It removes the datatype form (e.g. JOIN BINARY!) in favor of
letting MAKE BINARY! take over the role of the implicit
REDUCE.  Some open questions remain in terms of how sequences
will be handled, e.g. with MAKE PATH!, in terms of making sure
the dots and slashes line up.