UNGET, style prep for ^META void is isotope
This is some preparation for the codebase which would be in
support of the idea that the ^META of a pure void would be
a ~void~ isotope.  It means testing the Is_Meta_Of_Void() case
up front in several constructs.

This wouldn't affect FOR-BOTH or other users of the "friendly"
META that leaves pure void as pure void.  It would only affect
the more raw ^ operation, and draw more attention to places
where "pure voidness" was in play.

Further development is on a branch as this idea settles.