Remove vertical bar support from CHAIN
The syntax of `chain [:func1 | :func2]` arose when the | was
behaving as an "expression barrier".  Today that is done by
the COMMA! type, so having it supported by CHAIN has little
motivation.

Using COMMA! means that a plain REDUCE can be used by the
CHAIN wrapper over the native CHAIN*.