Remove lib modification in bootstrap-shim
The shim now makes the bootstrap executable COMPOSE act like
the modern COMPOSE (though it does not support /DEEP).  Before,
it would use a COMPOSE2 definition that was new...and could not
be seen normally by %make-spec.r in the extensions in a way
that would work in bootstrap.

Some code geared toward putting COMPOSE2 in lib was changed to
put COMPOSE and SPREAD in lib, which is now superfluous.  (It
actually causes damage when the APPEND to LIB is used to write
non-isotopic ACTION!s over top of isotopic ones in the new
model.)