Fix uncovered broken %test-parsing.r
There was an ELSE branch that wasn't running, but rather it
was evaluating to an ACTION! that was then being matched
against an input block.  This was detected by the isotope
branch.

So this puts the ELSE on a correct branch, which also made
it so that error delivery pointed out it's no longer based
on TEXT! (as tests now run on ordinarily LOAD-ed blocks, not
their own distinct "safe" loader).  So the line and column
information is not as granular--it only has the information
from the LINE OF the BLOCK!.