Change scanning rules for vertical bars
This makes it possible to put things with vertical bars in
them inside of PATH! and TUPLE!, and widens the legal range
of things that you can put inside the vertical bar escaped
notation.

It's no substitute for a scanner rewrite--but it's good enough
to move along to the next proposal.