Allow use of Symbol to make WORD! in feeds
With the coming concept of "action isotopes", it means that
the Lib() context will contain isotopes for most-if-not-all
actions.  Feeds will not take isotopes, so that means they
would have to be adjusted.

One strategy is to switch to using Canon() symbols instead,
which would put words that looked up to the actions.  This
has the benefit of not introducing an allocation that needs
to be freed to represent a quoted item in the feed, and also
makes the code more amenable to debugging.