First crack at generic isotopes
This introduces the notion of a "QUASI!" type, that replaces
the prior concept of BAD-WORD! with an entire class of values
that produce isotopes when evaluated.

What was formerly a BAD-WORD! with a null spelling is now a
"QUASI-BLANK!", that serves as the meta state for the none
isotope indicating an unset variable.

The META of a value is always either a QUASI!, a QUOTED!, or
a NULL.  This brings about a greater convergence between
quoting and meta-ing, though it does introduce some extra
unpacking.  But it also means that NULL is the only falsey
entry for META.

This is just a start with some rough edges, that manages to
run the tests without crashing.