Replace @void usages with void'
A new experiment is to try making ^META of void be null, while
the ^META of null would be '

To make this change easier, the current odd result of @void
for the meta of void is gotten through a word reference of
VOID'.  This is faster than saying `meta void` and can also
be used as the left hand side of a comparison without a group.

Adds NONE' and NULL' as well.