Eliminate @end ^META state, use VOID
There was a separate `<end>` state so you could tell the
difference between receiving a void argument and being at
the end of a feed of input.

When `<end>` was initially conceived, there would have been
no distinction--as the evaluator would have kept on feeding
if a void was received.

It's not worth contaminating the meta states for this--as
true END could be tested for on the frame feed itself if
one were interested in the origin of one's voids.