MATCHES for isotopic typematch in FIND/SWITCH
This addresses the longstanding ambiguity between trying to
do a FIND of a a datatype or typeset literally, vs finding
instances of that type.

They are now distinct:

    >> find block integer!  ; literally the integer datatype!

    >> find block matches integer!  ; integer instance

For a period of time, FIND and SELECT of DATATYPE! or TYPESET!
without a MATCHES are disabled, to raise an error and draw
attention to what should probably be a MATCHES case.

Also adds the ability to do a SWITCH on a MATCHES entry.