VOID propagates out of SET-WORD!, decays to NULL
If the evaluation step doesn't produce any output, we want
the variable to be set to NULL, but the voidness to propagate:

        >> 1 + 2 x: comment "hi"
        == 3

        >> x
        == null

Originally this would unset x, and propagate a none.  But once
isotope assignments were disallowed, that ruined

    x: y: if condition [...]

Where we want `x = y` after that.  Cases like

     return [# result]: ...

also presented a challenge for how not naming a variable
would want to preserve the value, so turning voids to nones
didn't feel good there.  The general `x: case [...]` being
able to leave X as NULL after no match was also a persuasion
point, along with the tight link between NULL and VOID by
triggering ELSE and linked by ^META state.  In the balance it
is simply too good to worry about the potentially misleading
state of X after the assignment.  So it was changed.