Limit Bootstrap Divider to strings
The bootstrap executable does not have sea of words, so if
you use words with the divider then that's making a bunch
of variables.  It's not all that inefficient, but the more
such things there are it makes it harder to do things like
see what the real variables in the global context are.

Limit it to strings just to be a little less aggressive in
the stress put on the old EXE.