Switch PAR() macro to be PARAM()
Originally the PARAM() macro was used for declaring parameters
and then ARG() would access it.

Then it became interesting to give access to the parameter
slots, so a PAR() macro was made.  Which has the advantage of
being three letters like ARG, but is otherwise ugly.

Since declaring parameters is done almost entirely by auto
generated code that no one sees, it makes more sense to use
something else there (DECLARE_PARAM) and then retake PARAM
to access the parameter slot.

This is also consistent with the macro redefinition of PARAM
in the action evaluator to refer to the current param, as that
parallels the ARG macro meaning the current arg.