Remove `<opt>` tag from most parser combinators
The idea of having parsers return definitional failure helps
make it work better with multi-return, because the variables
aren't overwritten on the left hand side...as well as other
good technical synergies.  So the only time a combinator
will return NULL is when it actually means it--as a successful
result (combinators like OPT would be an example, or anything
that calls arbitrary user code).

But previously every combinator had to say it could return
null.  This goes through and pares that down, correcting some
other type information in the process.