Improve %types.r, use PARSE on table
The types.r was processed in a tabular fashion.  Early Ren-C
attempted to generalize this for a table-oriented loader that
would have column names in a block in the first row that
defined the rest of the file...where using a generic function
you could enumerate that table as objects with the field
names as keys.

That is weaker than a PARSE-based dialect.  This does that,
and takes advantage of the freeform structure to introduce
TAG!s that mark ranges of typeset membership.