Eliminate SOMETHING? Native and macros
SOMETHING? is really just a test for if something is null
or blank.  This requires a particular interpretation of blank,
which the system is no longer taking.  Also since null cannot
be found in blocks, there are fewer cases where a value would
be either a blank or a null.