Automatically generate ordered type check macros
R3-Alpha had some brittle dependence on the relative ordering
of types in the type table.  The file %sys-ordered.h was
quarantined as the only place where macros exploiting that
knowledge could be defined.

This makes the jump to auto-generating these macros based on
added dialect information pointing out type ranges in the
%types.r file.