Use C99 variadics for optional args in continuations
The core has committed to C99 variadic macros (you can still
use the external API with C99 if you wish, though).

Given that, it's possible to implement functions that take
optional parameters.  This uses that ability for the
continue macros, so instead of having:

    return CONTINUE(OUT, block, with);
    return CONTINUE(OUT, block, END);

The difference is accomplished with:

    return CONTINUE(OUT, block, with);
    return CONTINUE(OUT, block);

In addition to being cleaner to be able to distinguish that
way, it also is helping to phase out the END idea--as that is
being removed from being a state cells can be in.