Finish eliminating "End Cells"
Once there was an idea that a special cell kind could be used
to signal an internal state that an end of input was reached.
This was unified with the idea of an array terminator, which
went through forms of compression over time (eventually being
eliminated and reduced down to a 2-byte signal used by the
external API only).

Other "end cell" uses were replaced by other ideas (notably the
output cell of a frame is initialized to stale voids, which
relies on two concepts that weren't around when the idea
Prior to the creation of stale states (which could be applied
to void) there was an idea that a special cell state would
be used to sometimes signal an end of input was reached.

This removes the last applications, mostly by having feeds use
a specialized signal that is compared by value to indicate
they are at the end.  This allows it to be distinguished from
the non-canon 2-byte rebEND (which would be more expensive to
test) and the nullptr state used by the API to represent a
nulled cell.