Allow DO to return VOID
There are some questions pending about the meaning of DO,
but having it be unwilling to return VOID is just confusing
matters of what it was asked.

It was also using a strange `return_non_void` function, which
needed to be deleted.

Requires use of DELEGATE_MAYBE_STALE, otherwise it will
overwrite the output with a non-stale void--forgetting any
previous cell (and being unable to assign to variables). This
likely needs a more holistic solution to stop it from happening.