Make NULL represent endable meta parameters
With the convention that ^META parameters are always either
QUOTED! or QUASI!, it opens up space for what other things
can mean.  This presents a chance for NULL to mean an
endable meta parameter.

As a result, the idea of the "~end~ isotope" could be
removed.  Also, the legality of representing voids in frame
cells now that they are null isotopes means that tricks
pertaining to the "~void~ isotope" during frame fulfillment
are not necessary, and plain void can be used.