Allow nulled internal cells in variadic feeds
The API implementation was paranoid about nulled cells leaking
to users, when they need to count on nullptr being the canon
form so it can be tested with C if and not need an API call.

However, it's easy enough to tell the difference between an
API handle and an internal cell...so there's no reason to
forbid internal nulleds.

There was some convolution to try and make it easier to not
pass nulls to the API, e.g. refinement tests being a nullptr
version of the argument.  But this just created noise (like
the need to write `did REF(arg)`.  This cleans that all up.