Improve system notion of "Erased/Poisoned" cells
At one point there was a decision to try and make it so
that you could initialize values into cells that had been
memset() with all 0s.  This could avoid a step of needing
to format the cell with NODE_FLAG_NODE and NODE_FLAG_CELL
before being able to initialize them.

This added the class INITABLE() to the existing READABLE()
and WRITABLE() states for cells.

It was a bit half-baked and didn't live up to its full
potential--partially due to some confusions about the nature
of the REB_0 state for void.  This became more sorted out,
and REB_0 was used for both NULL and VOID (with voids being
the isotopic form of null)...but as more of the zero bearing
states became legal it raised questions about how to leverage
those states illegally.

This revisits the question to try and get the most out of
the features and optimizations, while clarifying terminology
along the way.  It introduces the "erased" and "poisoned"
states, making it easier to work with cells that can be
neither read nor written--and clarifies the rules that make
it work.