Eliminate Is_End() and %sys-end.h
The END concept used to require significant explanation, when
it was a cell type used for unwritten frame outputs as well
as array termination.

It was whittled down into nonexistence--and is taken over by
the new features of Poison, Erased Cells, Staleness, and Voids.

The only remaining aspect doesn't fit in with cells at all,
but is just a 2-byte indicator of the end of a variadic feed.