Tighten optimizations/checks on "fresh" cells
It's desirable in general to notice when cells are being
accessed when they haven't been meaningfully initialized.
It's also desirable to be able to leverage optimizations
based on having the heart byte and quote byte of voids
both be zero, to be able to have a freshly "prep'd" cell
able to convert to a void using only one bitwise OR
operation.

This pushes on the definitions to get more pervasive
checks of the desired properties, geared around minimizing
the number of times the header needs to be masked or
assigned during the steps of construction.