Improve ergonomics of Append_Context()
Append_Context() was willing to take a possibly-quoted
ANY-WORD! instead of a symbol to expand with.  It would
then use the word to find the symbol, but also it bound the
word to the context at the position.

This splits that out into two functions that call an internal
implementation function--(which should have been done a long
time ago, but wasn't for some reason.)