Switch POKE to use non-^META frames
Having values be quoted complicated frame reuse, because the
methods were unquotifying the setval argument to POKE* that
then made it hard to tell whether it had been unquoted or
not.

Removing that makes things cleaner, and this permits the
POKE operation to protect the value cell so it can assure that
it is returning the value it got in.