Switch OPT_SYMID to use option(SymId)
Ren-C switched over to where dynamic symbols did not have
integer IDs.  This meant there was a danger in comparing the
symbol IDs of two words, because they might both be zero
indicating they have no integer ID... and compare the same.

An early technique for checking this made C++ classes for
optional symbol IDs as well as guaranteed non-zero ones.

Eventually the option() class was added, as a lightweight
version of std::optional that just worked for pointers.  But
it turns out that it's relatively easy to apply it to the
purpose of optional symbols.  So this does that, and it
cleans up the code as well as bringing more teeth to the
checking.