Allow evaluation into native LOCAL() cells
Historically you've been disallowed from doing evaluations
into frame cells, due to their public nature and a potential
for intersecting read/writes.  In particular, if a debugger
got access to a frame during a native evaluation into a cell,
it might expose a temporarily corrupt cell that was wiped.

But being able to do evaluations into native local cells is
very important with stackless, and shuffling evaluations
through SPARE every time is a pain.

This relaxes the rule to allow any LOCAL() cells in a native
to be used.