Make Gob_To_Array use stack
GOB! is now basically just a placeholder for techniques that might
be used by extension types.  It provides more code to test with
approaches and to help guide what future "real" extension
types might want.

Initialization of nulls was constrained so it could not be done
in a plain array cell.  This is good because it helps catch
at compile time cases where it shouldn't be done.  But it was
difficult for an array building function to express this by
writing into it without needing a temporary variable.

The right answer for code in such cases is probably to go ahead
and express the array construction via the stack.  This means
less code with more flexibility, at negligible cost in performance
(in this case it gets the array size exactly right, so a memory
savings that outweighs additional cost...especially for the
simplification in the code).