Make SPLIT-PATH a multi-return
There are several issues with the design of SPLIT-PATH.  This
adds some tests from the Red GitHub issue tracker to look at,
but doesn't resolve them at this time.

Rather, this just changes split path to use multiple return
values, so that its null states can be meaningfully reacted
to (and makes the callsites clearer).