Implement VOID-in-NULL-out Convention (core)
This reverts commit e0b38760f23480fdb1e1cbb08fce56ad6d65deac,
which took the BLANK-in-NULL out process and turned into
NULL-in-NULL-out by using definitional errors.

That mechanic was a good idea--so good that it has more
potential than opting out of parameters.  (If it is used for
opting out of parameters, it becomes conflated).

So this takes another angle, by using VOID-in-NULL-out.  The
specific properties of void's nothingness...that it cannot
come from a variable directly...makes it a good choice.