Transition bootstrap shim to BLANK! for NULL
The bootstrap EXE has a null non-valued type.  Unfortunately,
you cannot access a variable holding it through a plain word.
This accident of history makes it unpleasant to write code
that works in both the old and new EXE (and makes it unpleasant
to write code in the bootstrap EXE using NULL at all)

But what the bootstrap EXE's notion of "NULL" is actually maps
more closely to the current idea of VOID.  This reworks the
bootstrap shim so that plain BLANK! is used as the "NULL"
representation, and old NULL acts as void.

With some protections added to avoid accidental appends or
compositions of BLANK! with blocks, this can be rigged to
give an experience that is close to the target of today's EXE.