Use _ instead of the word NULL for initializations
To the right of a SET-WORD!, an _ looks good and helps make
null states stand out relative to other words:

    make object! [
       alpha: _
       beta: something
       gamma: _
    ]

It's also faster: the evaluator does not have to do a lookup to
see what NULL is set to.

Other places it looks better to use a WORD!, e.g. (return null)
looks more natural than (return _).  But it's a matter of taste.

This goes ahead and uses the underscore form for blanks for
assignments.