Clarify rebmake function calls with new APPLY
When Shixin wrote rebmake, there were functions that took a
lot of named parameters as refinements.  But there was no
APPLY operator, so the calls looked awkward.

The bootstrap shim now implements an APPLY that is like the
"new APPLY", so this replaces some of those awkward calls with
uses.