Use QUOTED! for FAIL's skippable BLAME parameter
In the early days of the @ type, it was thought that it would
be a good thing to use for pointing out the location that
an error was "at".

As it turns out, this is heavy and distracting, when there's
not really a lot of heft to what is intended.  This switches
it back to a quoted!, which is what it initially was.