Adjust tests for new TRAP rules
The previous "catch all" behavior for errors was dangerous, and
caused the ignoring of many errors that were misunderstood.

That has been moved to sys.util.rescue, with TRAP only reacting
to errors that come by contract from the immediate functions
that it calls.

This disrupted hundreds of tests, because patterns like this
would no longer work:

    assert error? trap [some code that makes an error]

Additionally that code frequently was not specific enough, and
would mask errors (even as simple as typos) vs. the error it
thought it was testing for.

Rather than put `sys.util.rescue` calls everywhere, this takes
advantage of the generic facility for recognizing errors by
ID in the test dialect.