Change type of splices from BLOCK! to GROUP!
Originally splices were canonized to isotopic blocks.  This
changes them to GROUP!, to make the canonization more obvious
(since most people would splice blocks, they'd see it reflected
that their type was not preserved).

It is also intended for aesthetic reasons of parentheses being
"softer boundaries", as well as fitting with some intended
purposes of ~()~ looking a bit like a 0 for using empty splices
as the product of a branch that produces void.

    >> if true []
    == ~()~

This sets it up to be something that--like a void--can be used
to represent empty intent to anything that takes a splice.

    >> append [a b c] if false []  ; append void
    == [a b c]

    >> append [a b c] if true []  ; append empty splice
    == [a b c]