Make void branches produce empty splices
This makes a taken branch that evaluates to void synthesize an
empty splice.

    >> if true []
    == ~()~  ; isotope

It replaces a prior behavior for identifying such cases, which
used the WORD! isotope of ~void~ to hint code outside the branch
at what happened.  Yet the choice to use a ~()~ isotope has an
immediate benefit, by being able to have the same intent as
void for many cases:

    >> append [a b c] if false [<a>]  ; IF returns void
    == [a b c]

    >> append [a b c] if true []  ; IF returns empty splice
    == [a b c]

While both intents "act void" for the purposes of things like
APPEND, only the true void will run an ELSE branch.  So this
maintains that coherence.