"null isotope" now BLANK! isotope, new NONE
Originally the idea of a "null isotope" was picking a quasiform
of a WORD! (because that was the only kind there was) to decay
to null.

With quasiforms for all types, it doesn't make sense to put
suspicion on the behavior of WORD! isotopes to think they will
always decay.  The BLANK! isotope can be that form, and it
makes sense that it decays to null now given the relationship
between NULL and BLANK!.

Further this makes the BLANK! go through an evaluation step
where it is a blank isotope before becoming a NULL.  This
can be leveraged by evaluations that want to know that they
were not pure null but from a blank (e.g. something like
DELIMIT could leverage that to turn blanks into spaces)

Because this retakes BLANK! isotopes from the "NONE" non
printing result of functions that don't want a result (such
as HELP, or PRINT itself), something else has to be used.
A forward-looking change turns this to be an empty isotopic
block, which if isotopic blocks are used for multi-return then
this would indicate no returns--hence could not be used for
an assignment with a SET-WORD!