Eliminate VAL_ARRAY_AT_HEAD()
There was a function in R3-Alpha called VAL_ARRAY_AT() that
would pick an element out of an array, but ignored the index.
This was error-prone, so Ren-C gave it the clearer name
of VAL_ARRAY_AT_HEAD() to emphasize that it wasn't accounting
for the index.

The goal was to get rid of instances, which had almost been
accomplished a long while ago...but not completely finished.
This goes ahead and pushes it through.

If you would like to ignore the index in an array, it's
possible to do so by extracting the array and using ARR_AT()
on the extracted array.