Drop WANTED() multi-return mechanic
The first implementation of multi-returns was done before
isotopic results existed, and so it used enfix to look left and
see if there was a SET-BLOCK!.  If so it did the multi-return
behavior.

This allowed an interesting twist: the function could actually
see how many return values were requested, and vary the behavior
based on that.  It was a trick used in things like TRANSCODE
to reduce how much you had to write in the call.

Then, the evaluator tried implementing automatic support for
multi-returns.  Yet internally it used much of the same concept,
of a relationship between the function and its left hand side.

Eventually the approach's limits showed the evaluator mechanic
needed to use something different: things like APPLY were not
easy to write, and anything that did an ENCLOSE of a multi-return
was thwarted by the relationship between the wrapped function
and its variables.  Even something as simple as putting an
expression in a GROUP! would break the composability.

So this required a new answer, based on putting multi-returns
in isotopic blocks.  A casualty of this is the ability of a
function to "sense" how many returns are requested.  But if one
wants a functionality in that vein, it's still possible to
use a "skin" for a multi-return that does it--if the downsides
are considered acceptable for the use case.